去除一部分Warning

+++ Xiaoer/External/GMGridView/GMGridViewLayoutStrategies.m
@@ -192,7 +192,7 @@
     int col = (int) (relativeLocation.x / (self.itemSize.width + self.itemSpacing)); 
     int row = (int) (relativeLocation.y / (self.itemSize.height + self.itemSpacing));
     
-    int position = col + row * self.numberOfItemsPerRow;
+    NSInteger position = col + row * self.numberOfItemsPerRow;
     
     if (position >= [self itemCount] || position < 0) 
     {
@@ -308,7 +308,7 @@
     int col = (int) (relativeLocation.x / (self.itemSize.width + self.itemSpacing)); 
     int row = (int) (relativeLocation.y / (self.itemSize.height + self.itemSpacing));
     
-    int position = row + col * self.numberOfItemsPerColumn;
+    NSInteger position = row + col * self.numberOfItemsPerColumn;
     
     if (position >= [self itemCount] || position < 0) 
     {
@@ -475,7 +475,7 @@
     int col = (int) (relativeLocation.x / (self.itemSize.width + self.itemSpacing)); 
     int row = (int) (relativeLocation.y / (self.itemSize.height + self.itemSpacing));
     
-    int position = [self positionForItemAtColumn:col row:row page:page];
+    NSInteger position = [self positionForItemAtColumn:col row:row page:page];
  
     if (position >= [self itemCount] || position < 0) 
     {
